+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 > Read detailed change logs in https://github.com/sakurawald/fuji/commits/dev/
 
 - [core] feature: improve the performance of search function.
+- [world] refactor: new implementation of `world` module.
+  - feature: improve the compatibility with other mods.
+  - feature: support `per-dimension difficulty`.
+  - feature: support `per-dimension time of day`. (Now you can `sleep` and set `spawn point` in `extra dimensions`.)
+  - feature: support `per-dimension weather`.